+++ tests/testthat/test-class-workbook-wrappers.R
@@ -403,6 +403,18 @@ test_that("wb_add_slicer() is a wrapper", {
   expect_wrapper("add_slicer", wb = wb, params = list(x = df, slicer = "vs", pivot_table = "pivot1"))
 })
 
+test_that("wb_add_timeline() is a wrapper", {
+  df <- data.frame(
+    date = seq(from = as.Date("2024-01-01"), length.out = 26, by = "month"),
+    amnt = sample(seq(100:150), 26, replace = TRUE)
+  )
+
+  wb <- wb_workbook()$add_worksheet()$add_data(x = df)
+  df <- wb_data(wb)
+  wb$add_pivot_table(x = df, data = "amnt", timeline = "date", pivot_table = "pivot1")
+  expect_wrapper("add_timeline", wb = wb, params = list(x = df, timeline = "date", pivot_table = "pivot1"))
+})
+
 test_that("wb_add_formula() is a wrapper", {
   wb <- wb_workbook()$add_worksheet(1)
   expect_wrapper("add_formula",    wb = wb, params = list(sheet = 1, x = "=TODAY()"))